Override HTTP_X_FORWARDED_PROTO header when merging Rack::Request env in replace_scheme
The reason for this is that latest (master) Rack::Request#scheme checks for
HTTP_X_FORWARDED_PROTO before rack.url_scheme. If the request was
proxied then this header is already set, which means that if the
original request was http, the resulting scheme is still http even
though we've set rack.url_scheme.