Add option for :mixed enforcement
There are cases where it is necessary to allow ssl on put or post but not get.
With restful routes this is harder since the actual route matches but the method changes.
The mixed option takes this into account and only forces strict on get and delete.

So the following:
  config.middleware.use Rack::SslEnforcer, :only => [/^\/posts\/(.+)\/edit/], :mixed => true

Above will allow PUT#post/:id to maintain the secure url while GET#post/:id will be forced to use http.