Explain how to install on sinatra/padrino
Adding RackSslEnforcer to sinatra or padrino is non-trivial.
The default session handler of both frameworks will place
the Rack::Session::Cookie middleware at the top, thus
RackSslEnforcer won't be able to see the cookies and operate
properly.

In order to fix it, you have to manually add Rack::Session::Cookie
as opposed to using enable :sessions.